mooveclient is now in commons folder
